Changed cast to be long literal

+++ src/main/java/com/e451/rest/domains/question/QuestionResponse.java
@@ -7,7 +7,7 @@ import java.util.List;
  */
 public class QuestionResponse {
     private List<Question> questions;
-    private Long paginationTotalElements = (long) 0;
+    private Long paginationTotalElements = 0L;
 
     public List<Question> getQuestions() {
         return questions;
